refactor: 역할별 파일 분할 리팩터링
