+++ index.html
@@ -35,5 +35,6 @@
       </main>
     </div>
   </body>
-  <script src="./js/DomHandler.js"></script>
+  <script type="module" src="./js/filter.js"></script>
+  <script type="module" src="./js/handleElement.js"></script>
 </html>
